requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
pyyaml
dvc